reorganize project code to adapt to new structure
