+++ app/(auth)/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/components/login/login-form"
+import { LoginForm } from "@/src/components/page/login/login-form";
 
 export default function LoginPage() {
   return (
@@ -7,5 +7,5 @@ export default function LoginPage() {
         <LoginForm />
       </div>
     </div>
-  )
+  );
 }
