+++ app/(protected)/dashboard/page.tsx
@@ -1,27 +1,29 @@
-import { createServerSupabaseClient } from '@/utils/supabase/server'
-import LogoutButton from '@/components/logout-button'
+import { createServerSupabaseClient } from "@/src/lib/supabase/server";
+import LogoutButton from "@/src/components/global/logout-button";
 
-export default async function DashboardPage(){
-     const supabase = await createServerSupabaseClient()  
-     const { data, error } = await supabase.auth.getUser()  
+export default async function DashboardPage() {
+  const supabase = await createServerSupabaseClient();
+  const { data, error } = await supabase.auth.getUser();
 
-     if (error) {
-         return (
-             <div className="flex items-center justify-center min-h-screen">
-                 <h1 className="text-4xl font-bold text-red-600">Error loading user data</h1>
-             </div>
-         )
-     }
-
+  if (error) {
     return (
-        <div className="flex items-center justify-center min-h-screen">
-            <div className="text-center">
-                <h1 className="text-4xl font-bold text-gray-800">Dashboard Page</h1>
-                <p className="mt-4">Welcome, {data?.user?.email || 'User'}</p>
-                <div className="mt-6">
-                    <LogoutButton />
-                </div>
-            </div>         
+      <div className="flex items-center justify-center min-h-screen">
+        <h1 className="text-4xl font-bold text-red-600">
+          Error loading user data
+        </h1>
+      </div>
+    );
+  }
+
+  return (
+    <div className="flex items-center justify-center min-h-screen">
+      <div className="text-center">
+        <h1 className="text-4xl font-bold text-gray-800">Dashboard Page</h1>
+        <p className="mt-4">Welcome, {data?.user?.email || "User"}</p>
+        <div className="mt-6">
+          <LogoutButton />
         </div>
-    )
+      </div>
+    </div>
+  );
 }
