--- components/logout-button.tsx
@@ -1,23 +1,0 @@
-'use client'
-
-import { createBrowserSupabaseClient } from '@/utils/supabase/client'
-import { useRouter } from 'next/navigation'
-
-export default function LogoutButton() {
-    const router = useRouter()
-
-    const handleLogout = async () => {
-        const supabase = createBrowserSupabaseClient()
-        await supabase.auth.signOut()
-        router.push('/')
-    }
-
-    return (
-        <button
-            onClick={handleLogout}
-            className="px-6 py-2 bg-red-600 text-white rounded-lg hover:bg-red-700 transition-colors"
-        >
-            Logout
-        </button>
-    )
-}
